requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pinecone-client==3.0.2
PyPDF2==3.0.1
python-dotenv==1.0.1
langchain==0.1.12
langchain-openai==0.0.8
openai==1.76.0
